Read description!
GitHub works in a branching system. When you open up your branch (click
the button next to tools in the top corner to select it), you will be
prompted to open up Explorer. Place any files you want to upload in the
folder, and then click the button to the left of the current branch to
synchronize and upload your new files.